+++ CHANGELOG.md
@@ -2,8 +2,18 @@
 
 ## Unreleased
 
+### Added
+
+* Handle multiple versions of the same plugin 
+* The 'cache' method now output all available versions for the same plugin
+* Support multiple arguments for 'install' command
+* Add '--upgrade' option for 'install' command 
+* Add '--pre' option for allowing  prerelease, development or experimental version
+  for 'install' and 'upgrade' command
+
 ### Fixed
 
+* Fix installation of experimental plugins
 * Improve output of 'update' command
 * Use templated remote definition in init as default
     - Allow forcing QGIS version